001_Scraping_Webharvy/004_Code/requirements.txt
google-cloud-bigquery
google-cloud-storage
python-dotenv
psutil
